requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow-cpu==2.10.0
# tensorflow-cpu>=2.15.0
matplotlib>=3.4.0
seaborn>=0.11.0
h5py>=3.1.0
openpyxl>=3.0.0
